refactor mgmp page
